Update the-100-year-computer.md
I made some minor changes. But I am worried that this is too terse to convey what we actually mean.

"Imagine a computer that was extremely valuable to you, that you’d pay thousands of dollars to recover, but you used *less* than the one you do today. "

This is confusing as hell. Are we saying that you would need to pay thousands of dollars to recover it? That that might happen? Why do you use it less than you use it today? Because it's just a better machine that needs less use? Are we saying that it's so valuable that you'd pay so much even for a device you rarely use?

So, general thought: we are trusting the reader to pick up on our hint too much.