UU Copy mostly finalized (needs a link audit) + basic bottom CTA on all pages
All of the copy elements of the site are up to parity with the author documents  — the only remaining thing to do is check that each link links appropriately and to the correct/intended location, as I had to populate each link myself based on the context of the writing.

Secondly, I've updated the CTA previously only located in "Project History, and it now lives on every UU page. I'd like to visually update this post-release, so consider this version an MVP.

The remaining task is to get the imagery up, bar any issues that come up in review.