Make clear that swap files on AWS are a bad idea
Adding a swap file on an Amazon EBS file system will allow Urbit to boot, but Urbit will <<behn>> permanently once enough memory is used  to cause that swap space to actually be accessed.  If swap is it to be used on AWS, it should be from an Instance Store volume.

This change mentions that swap doesn't work well on EBS, and points to Amazon's documentation on how to use swap on Instance Store.

This change still might not be enough to avoid user hassles, since the most common type of EC2 instances are unable to use Instance Store. But going into a long explanation about what EC2 instances will work with Urbit and which won't seems to be beyond the scope of this install guide.
updated for merge conflict

updated for merge conflict