Clarify and emphasise post-boot keyfile deletion
New users commonly run into problems when they try to run the initial boot command (with `-k`) for a subsequent boot - it's natural to think "this is how I started my ship the first time, and having shut it down I now want to start it again, so I'll run the same command". The existing recommendation to delete the keyfile doesn't make a note of *why* this recommendation is made, and so seems to often go ignored.