add full-stack walk-through
This guide walks through building a gall agent from scratch, building a
react front-end, putting together the desk, and publishing it. The
example app is called `Journal`, and is available in the new
[docs-examples](https://github.com/urbit/docs-examples) repo. The focus
of this guide is to show how everything fits together, with particular
focus on front-end integration. The guide is less interactive and more
demonstrative, though it contains ample instructive information.
Additional resource are listed for each part of the guide, so users can
get more information on anything covered.

The Gall Guide has been updated to reference this walk-through in a
couple of parts and link out to it at the end.

The example React app front-end is a bit too complicated to cover in a
blow-by-blow manner, so that part of the guide is more of a commentary
on the parts of the app that use the `http-api` library to communicate
with the ship.